+++ xlive/Blam/Engine/cache/cache_files.h
@@ -175,11 +175,11 @@ bool __cdecl cache_header_verify(s_cache_header* cache_header);
 
 uint32 __cdecl cache_file_align_read_size_to_cache_page(uint32 size);
 
-bool __cdecl cache_file_blocking_read(uint32 a1, uint32 cache_offset, uint32 read_size, void* out_buffer);
+bool __cdecl cache_file_blocking_read(intptr_t a1, uint32 cache_offset, uint32 read_size, void* out_buffer);
 
 bool __cdecl scenario_tags_load_internal(const char* scenario_path);
 
-datum tag_loaded(uint32 group_tag, const char* name);
+datum tag_loaded(int32 group_tag, const char* name);
 
 const char* tag_get_name(datum tag_name_index);
